+++ Core/NosSmooth.Game/Data/Inventory/InventoryItem.cs
@@ -1,0 +1,12 @@
+﻿//
+//  InventoryItem.cs
+//
+//  Copyright (c) František Boháček. All rights reserved.
+//  Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace NosSmooth.Game.Data.Inventory;
+
+/// <summary>
+/// Represents item in bag inventory of the character.
+/// </summary>
+public record InventoryItem();
