--- Core/NosSmooth.Game/Data/Inventory/Item.cs
@@ -1,12 +1,0 @@
-﻿// 
-// Item.cs
-// 
-// Copyright (c) Christofel authors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-namespace NosSmooth.Game.Data.Inventory;
-
-public class Item
-{
-    
-}
