+++ Core/NosSmooth.Game/Events/Core/IGameResponder.cs
@@ -4,10 +4,9 @@
 //  Copyright (c) František Boháček. All rights reserved.
 //  Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using OneOf.Types;
 using Remora.Results;
 
-namespace NosSmooth.Game.Events.Handlers;
+namespace NosSmooth.Game.Events.Core;
 
 /// <summary>
 /// Represents interface for classes that respond to <see cref="IGameEvent"/>.
@@ -18,7 +17,7 @@ public interface IGameResponder
 
 /// <summary>
 /// Represents interface for classes that respond to game events.
-/// Responds to <typeparamref name="TPacket"/>.
+/// Responds to <typeparamref name="TEvent"/>.
 /// </summary>
 /// <typeparam name="TEvent">The event type this responder responds to.</typeparam>
 public interface IGameResponder<TEvent> : IGameResponder
@@ -27,8 +26,8 @@ public interface IGameResponder<TEvent> : IGameResponder
     /// <summary>
     /// Respond to the given packet.
     /// </summary>
-    /// <param name="packet">The packet to respond to.</param>
+    /// <param name="gameEvent">The packet to respond to.</param>
     /// <param name="ct">The cancellation token for cancelling the operation.</param>
     /// <returns>A result that may or may not have succeeded.</returns>
-    public Task<Result> Respond(TEvent packet, CancellationToken ct = default);
+    public Task<Result> Respond(TEvent gameEvent, CancellationToken ct = default);
 }
